Add tests for query parser and visitors
